Allow only post method to delete tickets

+++ pages/api/delete-tickets/index.ts
@@ -3,7 +3,10 @@ import { withApiKeyAuthentication } from '../../../lib/middleware'
 import { unlink } from 'fs/promises'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  if (process.env.ALLOW_DEMO_FRONTEND !== 'allow') res.status(401)
+  if (process.env.ALLOW_DEMO_FRONTEND !== 'allow') return res.status(401)
+  if (req.method !== 'POST') {
+    return res.status(405).json({ Error: 'Cannot GET' })
+  }
   try {
     await unlink('./tickets.csv')
     res.status(200).send('Tickets deleted')
